test(seat-selection): compare heldSeats with int, not long (Sonar reliability gate)
assertThat(Integer).isEqualTo(long) compares dissimilar boxed types -
the one new-code bug failing PR #7's quality gate.

+++ backend/inventory-service/src/test/java/com/skybook/praveen/inventoryservice/integration/SeatHoldConcurrencyTest.java
@@ -216,7 +216,7 @@ class SeatHoldConcurrencyTest extends AbstractPostgresSpringBootTest {
         assertThat(failures).allSatisfy(failure -> assertThat(failure)
                 .isInstanceOfAny(SeatAlreadyHeldException.class, SeatNotAvailableException.class));
 
-        long successes = 2 - failures.size();
+        int successes = 2 - failures.size();
         assertThat(successes).isGreaterThanOrEqualTo(1);
         assertThat(activeHolds()).isEqualTo(successes);
